requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
colorama==0.4.6
Django==3.2.25
djangorestframework==3.15.1
exceptiongroup==1.2.2
importlib-metadata==6.7.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.2.0
pytest==7.4.4
pytz==2025.1
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
zipp==3.15.0